update int to decimal to account for price change to allow cents

+++ PierresVendors/Controllers/VendorsController.cs
@@ -40,7 +40,7 @@ namespace PierresVendors.Controllers
     }
 
     [HttpPost("/vendors/{vendorId}/orders")]
-    public ActionResult Create(int vendorId, string orderTitle, string orderInformation, string orderDate, int orderPrice)
+    public ActionResult Create(int vendorId, string orderTitle, string orderInformation, string orderDate, decimal orderPrice)
     {
       Dictionary<string, object> model = new Dictionary<string, object>();
       Vendor foundVendor = Vendor.Find(vendorId);
